Fix more krazy warnings.

+++ src/plugins/help/helpviewer_qwv.cpp
@@ -294,7 +294,7 @@ void HelpViewer::scaleUp()
 
 void HelpViewer::scaleDown()
 {
-    setTextSizeMultiplier(qMax(0.0, textSizeMultiplier() - 0.1));
+    setTextSizeMultiplier(qMax(qreal(0.0), textSizeMultiplier() - 0.1));
 }
 
 void HelpViewer::resetScale()
